third_party/WebKit/LayoutTests/inspector/elements/edit/edit-dom-actions-shadow-expected.txt

Tests that user can mutate author shadow DOM by means of elements panel.


Running: testSetUp

Running: testSetAuthorShadowDOMElementAttribute
==== before ====
- <div id="testSetAuthorShadowDOMElementAttribute">
    - #shadow-root (open)
          <div foo="attribute value" id="shadow-node-to-set-attribute"></div>
  </div>
==== after ====
- <div id="testSetAuthorShadowDOMElementAttribute">
    - #shadow-root (open)
          <div id="shadow-node-to-set-attribute" bar="edited attribute"></div>
  </div>

Running: testEditShadowDOMAsHTML
==== before ====
- <div id="testEditAuthorShadowDOMAsHTML">
    - #shadow-root (open)
          <div id="authorShadowDOMElement"></div>
  </div>
<div id="authorShadowDOMElement"></div>
==== after ====
- <div id="testEditAuthorShadowDOMAsHTML">
    - #shadow-root (open)
        - <span foo="shadow-span">
              <span id="inner-shadow-span">Shadow span contents</span>
          </span>
  </div>
